+++ spec/support/omniauth.rb
@@ -1,5 +1,4 @@
 OmniAuth.config.test_mode = true
-OmniAuth.config.path_prefix = "/admin/auth"
 OmniAuth.config.on_failure = proc { |env|
   OmniAuth::FailureEndpoint.new(env).redirect_to_failure
 }