+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog]
   form fields
 - Restrict access to `/admin` by IP
 - Fix student_loan_start_date validation error message
+- Fix OmniAuth failure path
 
 ## [Release 004] - 2019-09-04